Added the FR structure for stories and added 1001 Menus, French Talents shell
