+++ app/router/routes.js
@@ -88,6 +88,8 @@ import StoriesRockChoir from '../pages/stories/stories/rock-choir';
 import StoriesSpencerHockey from '../pages/stories/stories/spencer-hockey';
 import StoriesFoundationOfHearts from '../pages/stories/stories/foundation-of-hearts';
 import StoriesGreaterAnglia from '../pages/stories/stories/greater-anglia';
+import Stories1001Menus from '../pages/stories/stories/1001-menus';
+import StoriesTheFrenchTalents from '../pages/stories/stories/the-french-talents';
 
 export var homeRoute = 'home';
 
@@ -655,8 +657,11 @@ export const config = Immutable.fromJS([
     },
   ],
   [Stories, { name: 'stories' }, {
-      'en-GB': {
+      en: {
         path: '/stories',
+      },
+      fr: {
+        path: '/references',
       },
     },
   ],
@@ -705,6 +710,18 @@ export const config = Immutable.fromJS([
   [StoriesGreaterAnglia, { name: 'stories_greater_anglia', category: 'stories' }, {
       'en-GB': {
         path: '/stories/greater-anglia',
+      },
+    },
+  ],
+  [Stories1001Menus, { name: 'stories_1001_menus', category: 'stories' }, {
+      'fr': {
+        path: '/stories/1001-menus',
+      },
+    },
+  ],
+  [StoriesTheFrenchTalents, { name: 'stories_the_french_talents', category: 'stories' }, {
+      'fr': {
+        path: '/stories/the-french-talents',
       },
     },
   ],
